+++ opentelemetry-api/src/logs/logger.rs
@@ -1,6 +1,6 @@
-use std::borrow::Cow;
+use std::{borrow::Cow, sync::Arc};
 
-use crate::{logs::LogRecord, KeyValue};
+use crate::{logs::LogRecord, InstrumentationLibrary, KeyValue};
 
 /// The interface for emitting [`LogRecord`]s.
 pub trait Logger {
@@ -28,6 +28,37 @@ pub trait LoggerProvider {
         version: Option<Cow<'static, str>>,
         schema_url: Option<Cow<'static, str>>,
         attributes: Option<Vec<KeyValue>>,
+    ) -> Self::Logger {
+        self.library_logger(
+            Arc::new(InstrumentationLibrary::new(
+                name, version, schema_url, attributes,
+            )),
+        )
+    }
+
+    /// Returns a new versioned logger with the given instrumentation library.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use opentelemetry_api::{global, InstrumentationLibrary, logs::LoggerProvider};
+    ///
+    /// let provider = global::logger_provider();
+    ///
+    /// // logger used in applications/binaries
+    /// let logger = provider.logger("my_app");
+    /// // logger used in libraries/crates that optionally includes version and schema url
+    /// let library = std::sync::Arc::new(InstrumentationLibrary::new(
+    ///     env!("CARGO_PKG_NAME"),
+    ///     Some(env!("CARGO_PKG_VERSION")),
+    ///     Some("https://opentelemetry.io/schema/1.0.0"),
+    ///     None,
+    /// ));
+    /// let logger = provider.library_logger(library, true);
+    /// ```
+    fn library_logger(
+        &self,
+        library: Arc<InstrumentationLibrary>,
     ) -> Self::Logger;
 
     /// Returns a new logger with the given name.